fix(core): retention heatmap correctness and error display
- Rewrite v_analytics_retention_cohorts_weekly to emit the full (cohort_week × week_offset) grid via generate_series + LEFT JOIN, so cohorts with zero pebble activity still surface and W0 is fixed at 100% by definition
- Replace 604800 epoch math with date arithmetic ((next_week - cohort_week) / 7)
- Derive last_cohorts in get_retention_cohorts() from auth.users instead of the activity view, so silent cohorts appear in the last-8 window
- Split RetentionHeatmap.tsx: visualization stays, card wrapper moves to RetentionHeatmapCard.tsx (matches sibling ActiveUsersChart{,Card}.tsx)
- Wrap PostgrestError in Error in getRetentionCohorts so ErrorBlock renders the message instead of "[object Object]"
